Start of Temp Controller / Integration with Nest API / CAB Logs Implementation

+++ NestChecksAndBalances/Models/NestObject.cs
@@ -4,7 +4,7 @@ namespace NestChecksAndBalances.Models
 {
     public class NestObject
     {
-        public int humidity { get; set; }
+        public double humidity { get; set; }
         public string locale { get; set; }
         public string temperature_scale { get; set; }
         public bool is_using_emergency_heat { get; set; }
@@ -16,33 +16,33 @@ namespace NestChecksAndBalances.Models
         public string name { get; set; }
         public bool can_heat { get; set; }
         public bool can_cool { get; set; }
-        public int target_temperature_c { get; set; }
-        public int target_temperature_f { get; set; }
-        public int target_temperature_high_c { get; set; }
-        public int target_temperature_high_f { get; set; }
-        public int target_temperature_low_c { get; set; }
-        public int target_temperature_low_f { get; set; }
+        public double target_temperature_c { get; set; }
+        public double target_temperature_f { get; set; }
+        public double target_temperature_high_c { get; set; }
+        public double target_temperature_high_f { get; set; }
+        public double target_temperature_low_c { get; set; }
+        public double target_temperature_low_f { get; set; }
         public double ambient_temperature_c { get; set; }
-        public int ambient_temperature_f { get; set; }
+        public double ambient_temperature_f { get; set; }
         public double away_temperature_high_c { get; set; }
-        public int away_temperature_high_f { get; set; }
+        public double away_temperature_high_f { get; set; }
         public double away_temperature_low_c { get; set; }
-        public int away_temperature_low_f { get; set; }
+        public double away_temperature_low_f { get; set; }
         public double eco_temperature_high_c { get; set; }
-        public int eco_temperature_high_f { get; set; }
+        public double eco_temperature_high_f { get; set; }
         public double eco_temperature_low_c { get; set; }
-        public int eco_temperature_low_f { get; set; }
+        public double eco_temperature_low_f { get; set; }
         public bool is_locked { get; set; }
-        public int locked_temp_min_c { get; set; }
-        public int locked_temp_min_f { get; set; }
-        public int locked_temp_max_c { get; set; }
-        public int locked_temp_max_f { get; set; }
+        public double locked_temp_min_c { get; set; }
+        public double locked_temp_min_f { get; set; }
+        public double locked_temp_max_c { get; set; }
+        public double locked_temp_max_f { get; set; }
         public bool sunlight_correction_active { get; set; }
         public bool sunlight_correction_enabled { get; set; }
         public string structure_id { get; set; }
         public bool fan_timer_active { get; set; }
         public DateTime fan_timer_timeout { get; set; }
-        public int fan_timer_duration { get; set; }
+        public double fan_timer_duration { get; set; }
         public string previous_hvac_mode { get; set; }
         public string hvac_mode { get; set; }
         public string time_to_target { get; set; }
